+++ src/app/doctoralCenter/expert/(home)/page.jsx
@@ -1,8 +1,8 @@
 "use client";
 import Layout from "@/components/main-layout/Layout";
 
-import HomeGrid from "../_components/HomeGrid";
 import { path } from "../_constants/pathConstant";
+import HomeGrid from "../../(manager-expert)/_components/HomeGrid";
 import { sideMenu } from "../_constants/sideMenuConstants";
 
 export default function Page() {
