+++ src/app/doctoralCenter/expert/candidates/page.jsx
@@ -1,8 +1,8 @@
 "use client";
 import Layout from "@/components/main-layout/Layout";
-import { sideMenu } from "../_constants/sideMenuConstants";
 import { path } from "../_constants/pathConstant";
-import CandidatesGrid from "../_components/CandidatesGrid";
+import { sideMenu } from "../_constants/sideMenuConstants";
+import CandidatesGrid from "../../(manager-expert)/_components/CandidatesGrid";
 
 export default function Page() {
   return (
